test(evm): move RIPEMD-160 vectors to separate file
